tarkov-map-tracker/requirements.txt
streamlit>=1.30.0
folium>=0.15.0
gql[aiohttp]>=3.5.0
watchdog>=3.0.0
pyyaml>=6.0
requests>=2.31.0
streamlit-folium>=0.15.0
